Move excluded permissions control to separate component
