+++ src/views/auth/LoginComponent.vue
@@ -20,59 +20,7 @@
               <span class="user-name">{{ name }}</span>
             </div>
           </v-list-item>
-          <template v-if="disableablePermissions.length > 0 || hasSessionPermissions">
-            <v-list-item>
-              <v-list-item-title class="mb-1">{{
-                t('userSettings.permissions')
-              }}</v-list-item-title>
-              <template v-slot:append>
-                <v-btn
-                  v-if="permissionsChanged || hasSessionPermissions"
-                  color="primary"
-                  @click="
-                    permissionsChanged
-                      ? applyPermissionsAndClose()
-                      : resetPermissions()
-                  "
-                  size="small"
-                  variant="flat"
-                >
-                  {{
-                    hasSessionPermissions
-                      ? t('common.reset')
-                      : t('common.apply')
-                  }}
-                </v-btn>
-              </template>
-            </v-list-item>
-            <v-list-item>
-              <v-list density="compact" class="pa-0">
-                <v-list-item
-                  v-for="perm in disableablePermissions"
-                  :key="perm.id"
-                  class="pa-0"
-                >
-                  <v-checkbox
-                    density="compact"
-                    color="primary"
-                    :model-value="isPendingEnabled(perm.id)"
-                    @update:model-value="
-                      (val) => onPermissionChange(perm.id, val ?? true)
-                    "
-                    hide-details
-                    class="ml-3"
-                    :label="perm.id"
-                  >
-                    <template #label>
-                      <v-label class="text-subtitle-2 text-medium-emphasis">{{
-                        perm.id
-                      }}</v-label>
-                    </template>
-                  </v-checkbox>
-                </v-list-item>
-              </v-list>
-            </v-list-item>
-          </template>
+          <ExcludedPermissionsControl @close-menu="menuOpen = false" />
           <v-list-item @click="logout" v-if="hasUserManager">
             <v-list-item-title>{{ t('auth.signOut') }}</v-list-item-title>
           </v-list-item>
@@ -83,7 +31,8 @@
 </template>
 
 <script setup lang="ts">
-import { onMounted, ref, watch, computed } from 'vue'
+import { onMounted, ref, watch } from 'vue'
+import { useI18n } from 'vue-i18n'
 import { authenticationManager } from '../../services/authentication/AuthenticationManager.js'
 import { useRoute } from 'vue-router'
 import type { User } from 'oidc-client-ts'
@@ -92,8 +41,7 @@ import {
   initialsFromPreferredUserName,
 } from '@/lib/auth/initials.ts'
 
-import { useI18n } from 'vue-i18n'
-import usePermissionExcludes from '@/services/usePermissionExcludes'
+import ExcludedPermissionsControl from '@/components/permissions/ExcludedPermissionsControl.vue'
 
 const { t } = useI18n()
 const route = useRoute()
@@ -107,68 +55,7 @@ const name = ref(PLACEHOLDER_NAME)
 const user = ref<User | null>(null)
 const requiresLogin = ref(hasUserManager)
 
-const { permissions, isEnabled, togglePermission } = usePermissionExcludes()
-
-const permissionsChanged = ref(false)
-const pendingEnabled = ref<Record<string, boolean>>({})
 const menuOpen = ref(false)
-
-const STORAGE_KEY = 'v1-weboc-permission-excludes'
-const hasSessionPermissions = computed(() => {
-  try {
-    const stored = window.sessionStorage.getItem(STORAGE_KEY)
-    return stored && JSON.parse(stored).length > 0
-  } catch {
-    return false
-  }
-})
-
-const disableablePermissions = computed(() =>
-  permissions.value.filter((p) => p.assigned && isEnabled(p.id)),
-)
-function resetPermissions() {
-  window.sessionStorage.removeItem(STORAGE_KEY)
-  // Reset local state
-  Object.keys(pendingEnabled.value).forEach((key) => {
-    pendingEnabled.value[key] = true
-  })
-  permissionsChanged.value = false
-  menuOpen.value = false
-  window.location.reload()
-}
-
-watch(
-  permissions,
-  (perms) => {
-    perms.forEach((p) => {
-      if (!(p.id in pendingEnabled.value)) {
-        pendingEnabled.value[p.id] = isEnabled(p.id)
-      }
-    })
-  },
-  { immediate: true },
-)
-
-function isPendingEnabled(permId: string): boolean {
-  return pendingEnabled.value[permId] ?? isEnabled(permId)
-}
-
-function onPermissionChange(permissionId: string, included: boolean) {
-  pendingEnabled.value[permissionId] = included
-  permissionsChanged.value = permissions.value.some(
-    (p) => isPendingEnabled(p.id) !== isEnabled(p.id),
-  )
-}
-
-function applyPermissionsAndClose() {
-  permissions.value.forEach((p) => {
-    if (isPendingEnabled(p.id) !== isEnabled(p.id)) {
-      togglePermission(p.id, isPendingEnabled(p.id))
-    }
-  })
-  menuOpen.value = false
-  window.location.reload()
-}
 
 function setUser() {
   if (!hasUserManager) return
